+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+ - Remove unnecessary copy in `OpendalStore::set`
+
 ## [0.12.3] - 2024-03-07
 
 ### Added